+++ sei-ibc-go/modules/apps/transfer/keeper/msg_server.go
@@ -21,7 +21,7 @@ func (k Keeper) Transfer(goCtx context.Context, msg *types.MsgTransfer) (*types.
 
 	sequence, err := k.sendTransfer(
 		ctx, msg.SourcePort, msg.SourceChannel, msg.Token, sender, msg.Receiver, msg.TimeoutHeight, msg.TimeoutTimestamp,
-		msg.Metadata)
+		msg.Memo)
 	if err != nil {
 		return nil, err
 	}