+++ sei-ibc-go/modules/apps/transfer/keeper/relay_test.go
@@ -162,7 +162,7 @@ func (suite *KeeperTestSuite) TestOnRecvPacket() {
 		trace    types.DenomTrace
 		amount   sdk.Int
 		receiver string
-		metadata []byte
+		memo     string
 	)
 
 	testCases := []struct {
@@ -172,12 +172,12 @@ func (suite *KeeperTestSuite) TestOnRecvPacket() {
 		expPass      bool
 	}{
 		{"success receive on source chain", func() {}, true, true},
-		{"success receive on source chain with metadata", func() {
-			metadata = []byte("metadata")
+		{"success receive on source chain with memo", func() {
+			memo = "memo"
 		}, true, true},
 		{"success receive with coin from another chain as source", func() {}, false, true},
-		{"success receive with coin from another chain as source with metadata", func() {
-			metadata = []byte("metadata")
+		{"success receive with coin from another chain as source with memo", func() {
+			memo = "memo"
 		}, false, true},
 		{"empty coin", func() {
 			trace = types.DenomTrace{}
@@ -219,7 +219,7 @@ func (suite *KeeperTestSuite) TestOnRecvPacket() {
 			suite.coordinator.Setup(path)
 			receiver = suite.chainB.SenderAccount.GetAddress().String() // must be explicitly changed in malleate
 
-			metadata = []byte{}      // can be explicitly changed in malleate
+			memo = ""                // can be explicitly changed in malleate
 			amount = sdk.NewInt(100) // must be explicitly changed in malleate
 			seq := uint64(1)
 
@@ -246,14 +246,14 @@ func (suite *KeeperTestSuite) TestOnRecvPacket() {
 
 			// send coin from chainA to chainB
 			transferMsg := types.NewMsgTransfer(path.EndpointA.ChannelConfig.PortID, path.EndpointA.ChannelID, sdk.NewCoin(trace.IBCDenom(), amount), suite.chainA.SenderAccount.GetAddress().String(), receiver, clienttypes.NewHeight(0, 110), 0)
-			transferMsg.Metadata = metadata
+			transferMsg.Memo = memo
 			_, err := suite.chainA.SendMsgs(transferMsg)
 			suite.Require().NoError(err) // message committed
 
 			tc.malleate()
 
 			data := types.NewFungibleTokenPacketData(trace.GetFullDenomPath(), amount.String(), suite.chainA.SenderAccount.GetAddress().String(), receiver)
-			data.Metadata = metadata
+			data.Memo = memo
 			packet := channeltypes.NewPacket(data.GetBytes(), seq, path.EndpointA.ChannelConfig.PortID, path.EndpointA.ChannelID, path.EndpointB.ChannelConfig.PortID, path.EndpointB.ChannelID, clienttypes.NewHeight(0, 100), 0)
 
 			err = suite.chainB.GetSimApp().TransferKeeper.OnRecvPacket(suite.chainB.GetContext(), packet, data)